requirements.txt
Django==1.6.7
South==1.0
django-activeurl==0.1.8
django-braces==1.4.0
django-bootstrap3==4.11.0
django-guardian==1.2.4
django-userena==1.3.0
django-viewflow==0.5.0
django-tables2==0.15.0
django-taggit==0.12.2
django-sitetree==1.2.0
django-widget-tweaks==1.3
django-genericadmin==0.6.1
django-threadedcomments==0.9.0
django-activity-stream==0.5.1
django-tools==0.25.1
django-messages==0.5.1
easy-thumbnails==2.1
html2text==2014.9.8
pillow==2.6.0
psycopg2==2.5.4
six==1.7.3
sorl-thumbnail==11.12.1b
django-grappelli==2.5.5